+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.0.0
 
+- feat: add GuardedResult annotation
 - feat: apply proxy design pattern
 - feat: move builder to lib's root
 - feat: replace print to log